Reads list.txt and executes checkInput() on contents.
User can enter a single URL or a file name (only list.txt at the moment).
checkInput() will run for single URL as well as for each URL in
list.txt. Need to add error handling for nonexistent file and also for
keyboard interrupt - it crashes program instead of stopping it.